+++ test/controllers/moves_controller_test.rb
@@ -13,6 +13,11 @@ class MovesControllerTest < ActionDispatch::IntegrationTest
     assert_response :success
   end
 
+  test 'should get filtered index' do
+    get moves_url(tag: 'Cross Body')
+    assert_response :success
+  end
+
   test 'should get new' do
     get new_move_url
     assert_response :success