+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Added OpenSans font
 - **Moves Added**
   - Added ability to create, view, modify, and search through moves
+  - Moves can be filtered by clicking on tag names on the moves index
 - **Tags Added**
   - Moves can be tagged
   - Tags can be used to search for similar moves